// src/products/interfaces/products-response.ts
// Generated by https://quicktype.io

export interface IProductsResponse {
    id:          string;
    name:        string;
    slug:        string;
    image:       string;
    description: string;
    stock:       string;
    price:       number;
    deleted:     boolean;
    category:    Category;
}

export interface Category {
    id:          string;
    name:        string;
    slug:        string;
    description: string;
    image:       string;
}